Implementa el slider doble usando la librería materials de google

+++ app/src/main/java/com/example/pruebas/Invoice.java
@@ -3,13 +3,13 @@ package com.example.pruebas;
 // El objeto factura
 public class Invoice {
     private String descEstado;
-    private double importeOrdenacion;
+    private float importeOrdenacion;
     private String fecha;
 
     public String getDescEstado() {
         return descEstado;
     }
-    public double getImporteOrdenacion() {
+    public float getImporteOrdenacion() {
         return importeOrdenacion;
     }
     public String getFecha() {
